+++ Changelog.md
@@ -11,7 +11,8 @@ Initial release.
 
 ## Added
 
-* a `sphereplot(Manifolds.Sphere(2))` to start a plot with a sphere to put points on.
+* a `circleplot(Manifolds.Circle(𝔽))` for the real and complex circle representation
+* a `sphereplot(Manifolds.Sphere(n))` for `n=1,2` to start a plot with a sphere to put points on.
 * a `hyperboloidplot(Hyperbolic(n))` for `n=2,3` to start a plot in the hyperboloid model.
 * a `poincareballplot(Hyperbolic(n))` for `n=2,3` o start a plot in the Poincaré disc or ball model.
 * a `poincarehalfspaceplot(Hyperbolic(n))` for `n=2,3` to start a plot in the Poincaré hal plane or half space model.